[kaa]: don't crash offscreen vram size == 0
Okay, normally the functions
kaaPixmapUseScreen() and kaaPixmapUseMemory() should not
be used on pixmaps that have not been allocated with
offscreen vram support, i.e when offscreen vram size ==0.

Still the functions should not crash.

Added a log message during kaa initialisation to warn the user
when there is no offscreen vram.