[Xglamo] attempt to fix the "green line" bug.
Under some not well known circumstances, some 2D blits
result in some random green lights appearing on the destination
of the blit.
To whitness that, you have to for instance launch
the openmoko dialer, rotate the screen to right, rotate it back
and you will see some green lines in the text entry field of the
dialer.
To rotate the screen right and back to normal you have to type:
DISPLAY=:0 xrandr --output default --rotate right
DISPLAY=:0 xrandr --output default --rotate normal

This entry actually makes sure to not touch some 2D blitting
registers regions that are marked as reserved in the documentation.
To do that, the content of the registers are properly masked.

Also, the blitting code has been arranged to
not split the blitting between the "prepare"
and the actual "copy" phase. The blitting registers
are touched only in the actual "copy" phase.

If this appers to work out, we will need to do the same for
the solid fills as well.