[Xglamo] allow reseting of the command queue memory mappings
Unfortunately, it seems my green line problems are still not gone.
This commit should help in that respect.

Actually tracking the problem a bit further, I realized that
it appeared really only after going through the xrandr code path.
One think that happens after rotation or resize is that the offscreen
memory manager is reseted. Things that were in offscreen memory
are then "evitcted", and must be later pushed back into the new
reseted offscreen memory again.
So the memory mappings as seen by the command queue processor
must be updated.

This patch allows the command process to update his
offscreen memory mappings.